# internal/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "VATUSAs internal API for use by other internal VATUSA services.",
        "title": "VATUSA Internal",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "VATUSA Support",
            "url": "http://www.swagger.io/support",
            "email": "vatusa6@vatusa.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "host": "api.vatusa.local",
    "basePath": "/internal/v1",
    "paths": {
...
        "/user": {
            "get": {
                "description": "List users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "List users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/user.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
        "/user-flag": {
            "get": {
                "description": "List user flags",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-flag"
                ],
                "summary": "List user flags",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/user_flag.Response"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user flag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-flag"
                ],
                "summary": "Create a new user flag",
                "parameters": [
                    {
                        "description": "User Flag",
                        "name": "user_flag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user_flag.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/user_flag.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
        "/user-flag/{cid}": {
            "delete": {
                "description": "Delete a user flag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-flag"
                ],
                "summary": "Delete a user flag",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CID",
                        "name": "cid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Patch a user flag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-flag"
                ],
                "summary": "Patch a user flag",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CID",
                        "name": "cid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User Flag",
                        "name": "user_flag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user_flag.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user_flag.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
        "/user-flag/{id}": {
            "get": {
                "description": "Get a user flag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-flag"
                ],
                "summary": "Get a user flag",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User Flag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user_flag.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        },
...
        "/user/{cid}": {
            "get": {
                "description": "Get a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CID",
                        "name": "cid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CID",
                        "name": "cid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CID",
                        "name": "cid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Patch a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Patch a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CID",
                        "name": "cid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Role": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "Air Traffic Manager"
                },
                "role": {
                    "type": "string",
                    "example": "ATM"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                }
            }
        },
...
        "models.UserFlag": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "no_staff_log_entry_id": {
                    "type": "integer",
                    "example": 1
                },
                "no_staff_role": {
                    "type": "boolean",
                    "example": false
                },
                "no_training": {
                    "type": "boolean",
                    "example": false
                },
                "no_training_log_entry_id": {
                    "type": "integer",
                    "example": 1
                },
                "no_transferring": {
                    "type": "boolean",
                    "example": false
                },
                "no_transferring_log_entry_id": {
                    "type": "integer",
                    "example": 1
                },
                "no_visiting": {
                    "type": "boolean",
                    "example": false
                },
                "no_visiting_log_entry_id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                }
            }
        },
        "models.UserRole": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "facility_id": {
                    "type": "string",
                    "example": "ZDV"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                }
            }
        },
...
        "user.Request": {
            "type": "object",
            "required": [
                "cid",
                "controller_rating",
                "email",
                "first_name",
                "last_name",
                "pilot_rating"
            ],
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "controller_rating": {
                    "type": "integer",
                    "example": 1
                },
                "discord_id": {
                    "type": "string",
                    "example": "1234567890"
                },
                "email": {
                    "type": "string",
                    "example": "vatusa6@vatusa.net"
                },
                "first_name": {
                    "type": "string",
                    "example": "Raaj"
                },
                "last_cert_sync": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "last_name": {
                    "type": "string",
                    "example": "Patel"
                },
                "pilot_rating": {
                    "type": "integer",
                    "example": 1
                },
                "preferred_name": {
                    "type": "string",
                    "example": "Raaj"
                },
                "preferred_ois": {
                    "type": "string",
                    "example": "RP"
                }
            }
        },
        "user.Response": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "controller_rating": {
                    "type": "integer",
                    "example": 1
                },
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "discord_id": {
                    "type": "string",
                    "example": "1234567890"
                },
                "email": {
                    "type": "string",
                    "example": "vatusa6@vatusa.net"
                },
                "first_name": {
                    "type": "string",
                    "example": "Raaj"
                },
                "flags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserFlag"
                    }
                },
                "last_cert_sync": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "last_login": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "last_name": {
                    "type": "string",
                    "example": "Patel"
                },
                "pilot_rating": {
                    "type": "integer",
                    "example": 1
                },
                "pref_name_enabled": {
                    "type": "boolean",
                    "example": true
                },
                "preferred_name": {
                    "type": "string",
                    "example": "Raaj"
                },
                "preferred_ois": {
                    "type": "string",
                    "example": "RP"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserRole"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                }
            }
        },
        "user_flag.Request": {
            "type": "object",
            "required": [
                "cid"
            ],
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "no_staff_log_entry_id": {
                    "type": "integer",
                    "example": 1
                },
                "no_staff_role": {
                    "type": "boolean",
                    "example": false
                },
                "no_training": {
                    "type": "boolean",
                    "example": false
                },
                "no_training_log_entry_id": {
                    "type": "integer",
                    "example": 1
                },
                "no_transferring": {
                    "type": "boolean",
                    "example": false
                },
                "no_transferring_log_entry_id": {
                    "type": "integer"
                },
                "no_visiting": {
                    "type": "boolean",
                    "example": false
                },
                "no_visiting_log_entry_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "user_flag.Response": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "integer",
                    "example": 1293257
                },
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "no_staff_log_entry_id": {
                    "type": "integer",
                    "example": 1
                },
                "no_staff_role": {
                    "type": "boolean",
                    "example": false
                },
                "no_training": {
                    "type": "boolean",
                    "example": false
                },
                "no_training_log_entry_id": {
                    "type": "integer",
                    "example": 1
                },
                "no_transferring": {
                    "type": "boolean",
                    "example": false
                },
                "no_transferring_log_entry_id": {
                    "type": "integer",
                    "example": 1
                },
                "no_visiting": {
                    "type": "boolean",
                    "example": false
                },
                "no_visiting_log_entry_id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                }
            }
        },
...
        "utils.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "application-specific error code",
                    "type": "integer"
                },
                "error": {
                    "description": "application-level error message, for debugging",
                    "type": "string"
                },
                "status": {
                    "description": "user-level status message",
                    "type": "string"
                }
            }
        }
    }
}